Move worker info screen into debug info menu
No need to translate anything for debug info. Dunno what else will end up in that menu in the future.
